creation de la strecture de site web avec des groupes de root
