+++ app/(dashboard)/dashboard/layout.jsx
@@ -1,8 +1,5 @@
-
-import MainLayout from "@/components/MainLayout";
-
 import style from "./layout.module.css";
-import "./globals.css";
+import "@/app/globals.css";
 
 import { Inter } from "next/font/google";
 
@@ -15,15 +12,16 @@ const inter = Inter({
 });
 
 export const metadata = {
-  title: "ManageStock",
+  title: "ManageStock | Dashboard",
   description: "Generated by create next app",
 };
 
-export default function RootLayout({ children }) {
+export default function DashboardLayout({ children }) {
   return (
     <html lang="fr" className={`${inter.variable}`}>
       <body className={style.body}>
-        <MainLayout>{children}</MainLayout>
+        <div><h3>dashboard Layout</h3></div>
+        <main className={style.main}>{children}</main>
       </body>
     </html>
   );
