+++ app/(dashboard)/dashboard/page.jsx
@@ -1,5 +1,4 @@
 export default function Dashboard(){
-
     return(
         <>
         <h1>dashboard page</h1>
