+++ app/(manageStock)/layout.jsx
@@ -1,0 +1,32 @@
+import Header from "@/components/Header";
+import Footer from "@/components/Footer";
+
+import style from "./layout.module.css";
+import "@/app/globals.css";
+
+import { Inter } from "next/font/google";
+
+const inter = Inter({
+  subsets: ["latin"],
+  weight: ["400", "900", "700", "800", "500"],
+  style: ["italic", "normal"],
+  display: "swap",
+  variable: "--font-inter",
+});
+
+export const metadata = {
+  title: "ManageStock",
+  description: "Generated by create next app",
+};
+
+export default function ManageLayout({ children }) {
+  return (
+    <html lang="fr" className={`${inter.variable}`}>
+      <body className={style.body}>
+        <Header />
+        <main className={style.main}>{children}</main>
+        <Footer />
+      </body>
+    </html>
+  );
+}
