+++ app/(manageStock)/layout.module.css
