+++ app/(manageStock)/page.jsx
@@ -1,8 +1,5 @@
 import styles from "./page.module.css";
-import Form from "@/components/ContactezNous"
 
 export default function Home() {
-  return <>
-  <Form/>
-  </>;
+  return <></>;
 }
